+++ CHANGELOG.md
@@ -11,6 +11,7 @@ with the exception that 0.x versions can break between minor versions.
 - Autolink extension: Now supports configuration of different link types that
   should be recognized and converted to links. See `AutolinkExtension#builder`
   | Type    | Default? | Description                                               |
+  |---------|----------|-----------------------------------------------------------|
   | `URL`   | Yes      | URL with a protocol such as `https://example.com`         |
   | `EMAIL` | Yes      | Email address such as `foo@example.com`                   |
   | `WWW`   | No       | An address beginning with `www` such as `www.example.com` |