+++ CHANGELOG.md
@@ -17,7 +17,7 @@ with the exception that 0.x versions can break between minor versions.
   | `EMAIL` | Yes      | Email address such as `foo@example.com`                |
   | `WWW`   | Yes      | Address beginning with `www` such as `www.example.com` |
   
-  Note that This changes the behavior of `AutolinkExtension.create()` to now also
+  Note that this changes the behavior of `AutolinkExtension.create()` to now also
   include `WWW` links by default. To re-enable the previous behavior, use:
 
   ```java